+++ desktop-gpgs/src/de/golfgl/gdxgamesvcs/GameServiceSafeListener.java
@@ -9,7 +9,8 @@ import com.badlogic.gdx.Gdx;
  * 
  * @author mgsx
  * 
- * @todo pull-up this class to core package, could be usefull for any implementations.
+ * @todo pull-up this class to core package, could be usefull for any implementations. or
+ * better, force implementation to call listener in GLThread (postRunnable)
  *
  */
 public class GameServiceSafeListener implements IGameServiceListener {
